Conducive to splitting
